Make Block movable, align on size_t for size properties.
* Use size_t for dimensional properties to align with rest of code
  (and general container `.size()` types)
* Adopt MonoView for views
* Remove `Swap()` and `SetNumChannels()` methods to emphasize the
  implicit constness of the class (i.e. encourage setting dimensions
  only at construction time).

Bug: chromium:335805780

+++ modules/audio_processing/aec3/erle_estimator_unittest.cc
@@ -70,8 +70,8 @@ void FormFarendTimeFrame(Block* x) {
       11405,   15031.4, 14541.6, 19765.5, 18346.3, 19350.2, 3157.47, 18095.8,
       1743.68, 21328.2, 19727.5, 7295.16, 10332.4, 11055.5, 20107.4, 14708.4,
       12416.2, 16434,   2454.69, 9840.8,  6867.23, 1615.75, 6059.9,  8394.19};
-  for (int band = 0; band < x->NumBands(); ++band) {
-    for (int channel = 0; channel < x->NumChannels(); ++channel) {
+  for (size_t band = 0; band < x->NumBands(); ++band) {
+    for (size_t channel = 0; channel < x->NumChannels(); ++channel) {
       RTC_DCHECK_GE(kBlockSize, frame.size());
       std::copy(frame.begin(), frame.end(), x->begin(band, channel));
     }
@@ -108,8 +108,8 @@ void FormNearendFrame(
     std::array<float, kFftLengthBy2Plus1>* X2,
     rtc::ArrayView<std::array<float, kFftLengthBy2Plus1>> E2,
     rtc::ArrayView<std::array<float, kFftLengthBy2Plus1>> Y2) {
-  for (int band = 0; band < x->NumBands(); ++band) {
-    for (int ch = 0; ch < x->NumChannels(); ++ch) {
+  for (size_t band = 0; band < x->NumBands(); ++band) {
+    for (size_t ch = 0; ch < x->NumChannels(); ++ch) {
       std::fill(x->begin(band, ch), x->end(band, ch), 0.f);
     }
   }
